requirements.txt
python-can==4.3.1
pytest==9.0.2
pytest-mock==3.15.1